[SCMOD-8551] Updated group and packages
