+++ elastic-mapping-updater/pom.xml
@@ -21,7 +21,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>com.microfocus.caf.elastic</groupId>
+        <groupId>com.github.cafdataprocessing.elastic</groupId>
         <artifactId>elastic-mapping-updater-aggregator</artifactId>
         <version>1.0.0-SNAPSHOT</version>
     </parent>
