+++ elastic-mapping-updater/src/main/java/com/github/cafdataprocessing/elastic/tools/utils/FlatMapUtil.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.microfocus.caf.elastic.tools.utils;
+package com.github.cafdataprocessing.elastic.tools.utils;
 
 import java.util.AbstractMap;
 import java.util.LinkedHashMap;
