+++ pom.xml
@@ -25,7 +25,7 @@
         <version>2.1.0-129</version>
     </parent>
 
-    <groupId>com.microfocus.caf.elastic</groupId>
+    <groupId>com.github.cafdataprocessing.elastic</groupId>
     <artifactId>elastic-mapping-updater-aggregator</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <packaging>pom</packaging>
@@ -112,7 +112,7 @@
                 <version>25.0-jre</version>
             </dependency>
             <dependency>
-                <groupId>com.microfocus.caf.elastic</groupId>
+                <groupId>com.github.cafdataprocessing.elastic</groupId>
                 <artifactId>elastic-mapping-updater</artifactId>
                 <version>1.0.0-SNAPSHOT</version>
             </dependency>
